finished refactoring out component integration tests. modified input config of prompts to make when conditions unit testable
