+++ generators/app/project-configuration.ts
@@ -1,0 +1,58 @@
+// 'use strict';
+
+// /**
+//  * Describes the configuration settings for scaffolding a new project.
+//  *
+//  * @interface IProjectConfiguration
+//  */
+// interface IProjectConfiguration {
+//     /**
+//      * The name of the project that will be created.
+//      *
+//      * @type {string}
+//      * @memberOf IProjectConfiguration
+//      */
+//     name: string,
+
+//     /**
+//      * The description of the project that will be created.
+//      *
+//      * @type {string}
+//      * @memberOf IProjectConfiguration
+//      */
+//     description: string,
+
+//     /**
+//      * The type of project that will be created.
+//      *
+//      * @type {string}
+//      * @memberOf IProjectConfiguration
+//      */
+//     projectType: string,
+
+//     /**
+//      *
+//      *
+//      * @type {boolean}
+//      * @memberOf IProjectConfiguration
+//      */
+//     includeVSCode: boolean,
+
+//     /**
+//      *
+//      *
+//      * @type {string}
+//      * @memberOf IProjectConfiguration
+//      */
+//     author: string,
+
+//     /**
+//      *
+//      *
+//      * @type {boolean}
+//      * @memberOf IProjectConfiguration
+//      */
+//     installDependencies: boolean
+// }
+
+// export = IProjectConfiguration;
