+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "generator-swell",
-  "version": "0.11.1-alpha",
+  "version": "0.11.4-alpha",
   "description": "Yeoman generator for various types of Node, Angular, and Aurelia projects with TypeScript",
   "author": {
     "name": "swellaby",
