+++ test/component-integration/e2e/index-vsts-tests.ts
@@ -1,66 +1,59 @@
 'use strict';
 
-import yeomanAssert = require('yeoman-assert');
-import Chai = require('chai');
 import fs = require('fs');
 import helpers = require('yeoman-test');
 import path = require('path');
 import Sinon = require('sinon');
+import yeomanAssert = require('yeoman-assert');
 import YeomanGenerator = require('yeoman-generator');
-import yosay = require('yosay');
 
 import inputConfig = require('./../../../generators/app/input-config');
-import vsts = require('./../../../generators/app/vsts');
+import testHelpers = require('./../test-helpers');
 
 /**
  * Contains component integration tests from the index entry point to the functions in vsts.ts
  */
 suite('Index/VSTS Project Component Integration Tests:', () => {
     let sandbox: Sinon.SinonSandbox;
+    let gitInitCommandStub: Sinon.SinonStub;
+    let npmInstallCommandStub: Sinon.SinonStub;
+    let installDependenciesCommandStub: Sinon.SinonStub;
     const vstsCommonFiles = [
         'extension-icon.png',
         'OVERVIEW.md',
         'vss-extension.json'
     ];
 
-    const descriptionMessage = 'A new task to make a great platform even better';
-
     setup(() => {
-        sandbox = Sinon.sandbox.create();
+        sandbox = Sinon.sandbox.create()
+        gitInitCommandStub = testHelpers.createGitInitStub(sandbox);
+        npmInstallCommandStub = testHelpers.createNpmInstallStub(sandbox);
+        installDependenciesCommandStub = testHelpers.createDependenciesInstallStub(sandbox);
     });
 
     teardown(() => {
         sandbox.restore();
     });
 
-    suite('VSTS Task Option Tests:', () => {
-        const generatorRoot = path.join(__dirname, './../../../generators/app');
+    suite('VSTS Task Project Tests:', () => {
         const vstsAppName = 'vsts task';
         const appType = inputConfig.vstsTaskPromptValue;
         const appDescription = 'this is an awesome vsts task';
-        const yoDestinationPathFunctionName = 'destinationPath';
         const invalidParamsErrorMessage = 'Oh no! Encountered an unexpected error while trying to create a new VSTS ' +
-            'Task project :( The VSTS files were not added to the project.'
-        let gitInitCommandSpy: Sinon.SinonSpy;
-        let npmInstallCommandSpy: Sinon.SinonSpy;
-        let installDependenciesCommandSpy: Sinon.SinonSpy;
-        let consoleErrorSpy: Sinon.SinonSpy;
-
-        setup(() => {
-            gitInitCommandSpy = sandbox.spy(YeomanGenerator.prototype, 'spawnCommandSync').withArgs('git', ['init', '--quiet']);
-            npmInstallCommandSpy = sandbox.spy(YeomanGenerator.prototype, 'npmInstall');
-            installDependenciesCommandSpy = sandbox.spy(YeomanGenerator.prototype, 'installDependencies');
-            consoleErrorSpy = sandbox.spy(console, 'error');
-        });
+            'Task project :( The VSTS files were not added to the project.';
 
         suiteSetup(() => {
-            return helpers.run(generatorRoot)
+            return helpers.run(testHelpers.generatorRoot)
                 .withPrompts({
                     appName: vstsAppName,
                     description: appDescription,
                     type: appType
                 })
                 .toPromise();
+        });
+
+        test('Should create all the correct boilerplate files when the VSTS option is selected', () => {
+            yeomanAssert.file(testHelpers.boilerplateFiles);
         });
 
         test('Should contain all of the common VSTS files', () => {
@@ -79,12 +72,12 @@ suite('Index/VSTS Project Component Integration Tests:', () => {
         });
 
         test('Should inject the App Name into the README.md file when the VSTS option is selected', () => {
-            yeomanAssert.fileContent('README.md', '# ' + vstsAppName);
+            yeomanAssert.fileContent(testHelpers.readmeFileName, '# ' + vstsAppName);
         });
 
         test('Should create and scaffold into a new directory if the specified app name differs from the'
             + 'current directory with the VSTS option', (done) => {
-                helpers.run(generatorRoot)
+                helpers.run(testHelpers.generatorRoot)
                     .withPrompts({
                         appName: vstsAppName,
                         description: 'my test',
@@ -96,15 +89,15 @@ suite('Index/VSTS Project Component Integration Tests:', () => {
                         yeomanAssert.equal(path.resolve(process.cwd()), path.join(dir, vstsAppName));
                         done();
                     });
-            });
+        });
 
         test('Should scaffold into the current directory when the specified app name matches the current'
             + 'directory name with the VSTS option', (done) => {
-                sandbox.stub(YeomanGenerator.prototype, yoDestinationPathFunctionName, () => {
+                sandbox.stub(YeomanGenerator.prototype, testHelpers.yoDestinationPathFunctionName).callsFake(() => {
                     return path.join(process.cwd(), vstsAppName);
                 });
 
-                helpers.run(generatorRoot)
+                helpers.run(testHelpers.generatorRoot)
                     .withPrompts({
                         appName: vstsAppName,
                         description: appDescription,
@@ -117,10 +110,10 @@ suite('Index/VSTS Project Component Integration Tests:', () => {
                         yeomanAssert.noFile(path.join(process.cwd(), vstsAppName));
                         done();
                     });
-            });
+        });
 
         test('Should install dependencies if user confirms with the VSTS option selected', (done) => {
-            helpers.run(generatorRoot)
+            helpers.run(testHelpers.generatorRoot)
                 .withPrompts({
                     appName: 'name',
                     description: appDescription,
@@ -129,14 +122,14 @@ suite('Index/VSTS Project Component Integration Tests:', () => {
                 })
                 .toPromise()
                 .then(() => {
-                    yeomanAssert.deepEqual(npmInstallCommandSpy.called, true);
-                    yeomanAssert.deepEqual(installDependenciesCommandSpy.called, false);
+                    yeomanAssert.deepEqual(npmInstallCommandStub.called, true);
+                    yeomanAssert.deepEqual(installDependenciesCommandStub.called, false);
                     done();
                 });
         });
 
         test('Should not install dependencies if user declines with the VSTS option selected', (done) => {
-            helpers.run(generatorRoot)
+            helpers.run(testHelpers.generatorRoot)
                 .withPrompts({
                     appName: vstsAppName,
                     description: appDescription,
@@ -145,8 +138,32 @@ suite('Index/VSTS Project Component Integration Tests:', () => {
                 })
                 .toPromise()
                 .then(() => {
-                    yeomanAssert.deepEqual(npmInstallCommandSpy.called, false);
-                    yeomanAssert.deepEqual(installDependenciesCommandSpy.called, false);
+                    yeomanAssert.deepEqual(npmInstallCommandStub.called, false);
+                    yeomanAssert.deepEqual(installDependenciesCommandStub.called, false);
+                    done();
+                }
+            );
+        });
+
+        test('Should create all the correct VS Code files when the vscode option is selected', () => {
+            yeomanAssert.file(testHelpers.vsCodeFiles);
+        });
+
+        test('Should set VS Code debug program correctly for VSTS app', () => {
+            yeomanAssert.fileContent('.vscode/launch.json', '"program": "${file}"');
+        });
+
+        test('Should not add VS Code files if the user declines the vscode option', (done) => {
+            helpers.run(testHelpers.generatorRoot)
+                .withPrompts({
+                    appName: 'vsts',
+                    description: appDescription,
+                    type: inputConfig.vstsTaskPromptValue,
+                    vscode: false
+                })
+                .toPromise()
+                .then(() => {
+                    yeomanAssert.noFile(testHelpers.vsCodeFiles);
                     done();
                 });
         });
