+++ test/component-integration/test-helpers.ts
@@ -1,8 +1,12 @@
 'use strict';
 
 import path = require('path');
+import Sinon = require('sinon');
+import YeomanGenerator = require('yeoman-generator');
 
 export const generatorRoot = path.join(__dirname, './../../generators/app');
+export const readmeFileName = 'README.md';
+export const yoDestinationPathFunctionName = 'destinationPath';
 
 export const boilerplateFiles = [
     '.eslintignore',
@@ -22,3 +26,40 @@ export const boilerplateFiles = [
     './build/tasks/lint.js',
     './build/tasks/test.js'
 ];
+
+export const vsCodeFiles = [
+    '.vscode/tasks.json',
+    '.vscode/extensions.json',
+    '.vscode/launch.json'
+];
+
+/**
+ * Creates the Sinon stub of the spawnSync method to init a git repo on Yeoman Generator instances.
+ *
+ * @param sandbox
+ */
+export const createGitInitStub = (sandbox: Sinon.SinonSandbox): Sinon.SinonStub => {
+    return sandbox.stub(YeomanGenerator.prototype, 'spawnCommandSync').withArgs('git', ['init', '--quiet']);
+};
+
+/**
+ * Creates the Sinon stub of the npmInstall method on Yeoman Generator instances.
+ *
+ * @param sandbox
+ */
+export const createNpmInstallStub = (sandbox: Sinon.SinonSandbox): Sinon.SinonStub => {
+    return sandbox.stub(YeomanGenerator.prototype, 'npmInstall');
+};
+
+/**
+ * Creates the Sinon stub of the installDependencies method on Yeoman Generator instances.
+ *
+ * @param sandbox
+ */
+export const createDependenciesInstallStub = (sandbox: Sinon.SinonSandbox): Sinon.SinonStub => {
+    return sandbox.stub(YeomanGenerator.prototype, 'installDependencies');
+};
+
+export const createConsoleErrorStub = (sandbox: Sinon.SinonSandbox): Sinon.SinonStub => {
+    return sandbox.stub(console, 'error');
+};
